requirements.txt
dnslib==0.9.25
typing_extensions==4.12.2
redis==5.2.1
python_dotenv=1.0.1
os=1.1.1